sample_receiver/requirements.txt
asgiref==3.4.1
Django==2.2
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
model-bakery==1.3.2
pytest==6.2.4
pytest-django
pytz==2021.1
sqlparse==0.4.1
